+++ example/maze/index.dart
@@ -1,0 +1,15 @@
+library maze_gen;
+
+import 'dart:html';
+import 'package:boarding/grid.dart';
+import 'package:boarding/util.dart';
+import 'package:boarding/boarding.dart';
+
+part 'model/grid.dart';
+part 'view/board.dart';
+
+main() {
+  var canvas = querySelector('#canvas');
+  var table = new Table(new Area(canvas.width, canvas.height), new Size(30, 30));
+  new Board(canvas, new MazeGrid(table)).draw();
+}